ML/Backend/requirements.txt
Flask==2.1.1
flask-cors==3.0.10
whisper==0.1.1
fpdf==1.7.2
requests==2.26.0
google-generativeai==0.1.0
